requirements.txt
autopep8==1.5.4
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
python-dotenv==0.14.0
sentry-sdk==0.18.0
toml==0.10.1
urllib3==1.25.10pip3 freeze > requirements.txt
Werkzeug==1.0.1
